better control of key iteration and handling
by double iterating on ID 0, will always have a problem unmarshalling.

instead increment the job id 1 over the sequence key so will never be 0. The purpose
of no zeroes is so can tell if during addition to the db
the job ID can be checked that it was incremented,
as zero is the default value when the struct
is created